+++ src/docs/asciidoc/api/member/login.adoc
@@ -1,0 +1,14 @@
+[[login]]
+== 로그인
+
+=== 로그인 요청
+
+include::{snippets}/login/http-request.adoc[]
+include::{snippets}/login/request-fields.adoc[]
+
+=== 로그인 응답
+
+include::{snippets}/login/http-response.adoc[]
+include::{snippets}/login/response-fields.adoc[]
+
+[[login]]
