+++ src/docs/asciidoc/api/member/name-duplicate.adoc
@@ -1,0 +1,14 @@
+[[name-duplicate]]
+== 이름(닉네임) 중복 체크
+
+=== 이름(닉네임) 중복 체크  요청
+
+include::{snippets}/name-duplicate/http-request.adoc[]
+include::{snippets}/name-duplicate/path-parameters.adoc[]
+
+=== 이름(닉네임) 중복 체크  응답
+
+include::{snippets}/name-duplicate/http-response.adoc[]
+include::{snippets}/name-duplicate/response-fields.adoc[]
+
+[[name-duplicate]]
