+++ src/main/java/com/petlink/member/dto/response/NameCheckResponse.java
@@ -1,0 +1,12 @@
+package com.petlink.member.dto.response;
+
+import lombok.Getter;
+
+@Getter
+public class NameCheckResponse {
+	Boolean isOk;
+
+	public NameCheckResponse() {
+		this.isOk = Boolean.TRUE;
+	}
+}
