+++ src/main/java/com/petlink/member/exception/MemberExceptionCode.java
@@ -11,6 +11,7 @@ public enum MemberExceptionCode {
 
 	ALREADY_REGISTERED_MEMBER(HttpStatus.BAD_REQUEST, "이미 등록된 회원입니다."),
 	NOT_FOUND_MEMBER_EXCEPTION(HttpStatus.NOT_FOUND, "회원을 찾을 수 없습니다."),
+	EMAIL_NOT_VALID(HttpStatus.BAD_REQUEST, "이메일 형식이 올바르지 않습니다."),
 	NOT_MATCHED_INFOMATION(HttpStatus.UNAUTHORIZED, "정보가 일치하지 않습니다."),
 	ALREADY_USED_NAME(HttpStatus.CONFLICT, "이미 사용중인 이름입니다.");
