+++ src/test/java/com/petlink/member/controller/MemberControllerTest.java
@@ -2,8 +2,13 @@ package com.petlink.member.controller;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
-
-import java.util.Objects;
+import static org.springframework.http.MediaType.*;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.*;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
+import static org.springframework.restdocs.payload.PayloadDocumentation.*;
+import static org.springframework.restdocs.request.RequestDocumentation.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -11,19 +16,19 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
+import org.springframework.restdocs.payload.JsonFieldType;
 
+import com.petlink.RestDocsSupport;
 import com.petlink.member.dto.request.SignUpRequestDto;
 import com.petlink.member.dto.response.MemberInfoResponseDto;
+import com.petlink.member.dto.response.NameCheckResponse;
 import com.petlink.member.exception.MemberException;
 import com.petlink.member.exception.MemberExceptionCode;
 import com.petlink.member.service.MemberService;
 
 @ExtendWith(MockitoExtension.class)
-@DisplayName("MemberController 테스트")
-class MemberControllerTest {
+class MemberControllerTest extends RestDocsSupport {
 
 	@InjectMocks
 	private MemberController memberController;
@@ -31,38 +36,15 @@ class MemberControllerTest {
 	@Mock
 	private MemberService memberService;
 
-	@Mock
-	private PasswordEncoder passwordEncoder;
-
-	@Test
-	@DisplayName("회원가입을 진행할 수 있다.")
-	void signUpTest() {
-		SignUpRequestDto signUpRequestDto = SignUpRequestDto.builder()
-			.name("TestName")
-			.email("test@example.com")
-			.password("password")
-			.tel("1234567890")
-			.zipCode("12345")
-			.address("TestAddress")
-			.detailAddress("TestDetailAddress")
-			.build();
-
-		signUpRequestDto.encodingPassword(passwordEncoder.encode(signUpRequestDto.getPassword()));
-		MemberInfoResponseDto memberInfoResponseDto = MemberInfoResponseDto.of(signUpRequestDto.toEntity());
-
-		when(memberService.signUp(any(SignUpRequestDto.class))).thenReturn(memberInfoResponseDto);
-		ResponseEntity<MemberInfoResponseDto> responseEntity = memberController.signUp(signUpRequestDto);
-
-		verify(memberService, times(1)).signUp(signUpRequestDto);
-		assertEquals(HttpStatus.CREATED, responseEntity.getStatusCode());
-		assertEquals(memberInfoResponseDto.getEmail(), Objects.requireNonNull(responseEntity.getBody()).getEmail());
+	@Override
+	protected Object initController() {
+		return new MemberController(memberService);
 	}
 
 	@Test
-	@DisplayName("올바르지 않은 이메일 형식은 회원가입을 진행할 수 없다.")
-	void signUpTest_invalidEmail() {
-
-		SignUpRequestDto signUpRequestDto = SignUpRequestDto.builder()
+	@DisplayName("회원가입을 진행할 수 있다.")
+	void signUpTest() throws Exception {
+		SignUpRequestDto request = SignUpRequestDto.builder()
 			.name("TestName")
 			.email("test@example.com")
 			.password("password")
@@ -71,22 +53,77 @@ class MemberControllerTest {
 			.address("TestAddress")
 			.detailAddress("TestDetailAddress")
 			.build();
-		signUpRequestDto.encodingPassword(passwordEncoder.encode(signUpRequestDto.getPassword()));
 
-		when(memberService.signUp(any(SignUpRequestDto.class)))
-			.thenThrow(new IllegalArgumentException("이메일 형식이 올바르지 않습니다."));
+		MemberInfoResponseDto response = MemberInfoResponseDto.builder()
+			.id(1L)
+			.name(request.getName())
+			.email(request.getEmail())
+			.build();
 
-		assertThrows(IllegalArgumentException.class, () -> memberController.signUp(signUpRequestDto));
+		when(memberService.signUp(any(SignUpRequestDto.class))).thenReturn(response);
 
-		verify(memberService, times(1)).signUp(signUpRequestDto);
+		mockMvc.perform(
+				post("/members/signup")
+					.content(objectMapper.writeValueAsString(request))
+					.contentType(APPLICATION_JSON)
+			)
+			.andExpect(status().isCreated())
+			.andExpect(jsonPath("id").value(response.getId()))
+			.andExpect(status().isCreated())
+			.andDo(document("sign-up",
+				preprocessRequest(prettyPrint()),
+				preprocessResponse(prettyPrint()),
+				requestFields(
+					fieldWithPath("name").type(JsonFieldType.STRING).description("이름(닉네임)"),
+					fieldWithPath("email").type(JsonFieldType.STRING).description("이메일"),
+					fieldWithPath("password").type(JsonFieldType.STRING).description("비밀번호"),
+					fieldWithPath("tel").type(JsonFieldType.STRING).description("전화번호"),
+					fieldWithPath("zipCode").type(JsonFieldType.STRING).description("우편번호"),
+					fieldWithPath("address").type(JsonFieldType.STRING).description("주소"),
+					fieldWithPath("detailAddress").optional().type(JsonFieldType.STRING).description("상세주소")
+				),
+				responseFields(
+					fieldWithPath("id").type(JsonFieldType.NUMBER).description("회원번호"),
+					fieldWithPath("name").type(JsonFieldType.STRING).description("이름(닉네임)"),
+					fieldWithPath("email").type(JsonFieldType.STRING).description("이메일")
+				)
+			));
+
+		verify(memberService, times(1)).signUp(any(SignUpRequestDto.class));
+	}
+
+	@Test
+	@DisplayName("이름(닉네임)을 중복 검증 할 수 있다.")
+	void checkNameTest() throws Exception {
+		String testName = "TestName";
+		NameCheckResponse response = new NameCheckResponse();
+
+		when(memberService.isNameDuplicated(testName)).thenReturn(response);
+
+		mockMvc.perform(
+				RestDocumentationRequestBuilders.get("/members/duplicate/{name}", testName)
+			)
+			.andExpect(status().isOk())
+			.andExpect(jsonPath("isOk").value(response.getIsOk()))
+			.andDo(document("name-duplicate",
+				preprocessRequest(prettyPrint()),
+				preprocessResponse(prettyPrint()),
+				pathParameters(
+					parameterWithName("name").description("이름(닉네임)")
+				),
+				responseFields(
+					fieldWithPath("isOk").type(JsonFieldType.BOOLEAN).description("중복 여부")
+				)
+			));
+
+		verify(memberService, times(1)).isNameDuplicated(testName);
 	}
 
 	@Test
 	@DisplayName("중복된 이름(닉네임)은 회원가입 할 수 없다.")
-	void checkNameTest() {
+	void checkNameFailTest() {
 
 		String testName = "TestName";
-
 		when(memberService.isNameDuplicated(testName)).thenThrow(
 			new MemberException(MemberExceptionCode.ALREADY_USED_NAME));
